calculator interface changed,2

+++ SEGruppo8/src/main/java/it/unisa/SoftwreEngineering/Gruppo8/Calculator.java
@@ -75,7 +75,7 @@ public class Calculator {
     /**
      *
      */
-    public void insert(int real ,int img){
+    public void insert(double real ,double img){
         throw new UnsupportedOperationException("Not supported yet.");
     } 
 }
